Refactor : 패키지 구조 변경, 네이밍 변경
